+++ .travis.yml
@@ -11,7 +11,7 @@ before_install:
   - pip install -U pytest
   - pip install codecov
 script:
-    - cd ICC/tests/
+    - cd icc/tests/
       - python icc_tests.py
 
 after_success:
